# docs/swagger.yaml
definitions:
  chaincode.invokeCCRequest:
    properties:
      cc_name:
        type: string
      channel_name:
        type: string
      function:
        type: string
    type: object
  lifecycle.approveCCRequest:
    properties:
      cc_name:
        type: string
      cc_sequence:
        type: integer
      cc_version:
        type: string
      channel_name:
        type: string
      package_ID:
        type: string
    type: object
  lifecycle.approvedChaincodeResponse:
    properties:
      endorsement_plugin:
        type: string
      init_required:
        type: boolean
      package_ID:
        type: string
      sequence:
        type: integer
      validation_plugin:
        type: string
      version:
        type: string
    type: object
...
  lifecycle.commitCCRequest:
    properties:
      cc_name:
        type: string
      cc_sequence:
        type: integer
      cc_version:
        type: string
      channel_name:
        type: string
    type: object
  lifecycle.committedChaincodeResponse:
    properties:
      chaincode:
        type: string
      channel:
        type: string
      endorsement_plugin:
        type: string
      sequence:
        type: string
      validation_plugin:
        type: string
      version:
        type: string
    type: object
  lifecycle.currentAdmin:
    properties:
      admin:
        type: string
    type: object
...
  lifecycle.packageCCRequest:
    properties:
      cc_source_name:
        type: string
      label:
        type: string
      language:
        type: string
      package_name:
        type: string
    type: object
  lifecycle.queryRequest:
    properties:
      cc_name:
        type: string
      channel_name:
        type: string
    type: object
...
info:
  contact:
    name: arogya.Gurkha
    url: https://github.com/arogyaGurkha
  description: API to run fabric binaries
  title: Hyperledger Fabric Gurkhaman API
  version: "0.1"
paths:
  /fabric/chaincode/invoke:
    post:
      consumes:
      - application/json
      description: '`peer chaincode invoke` is executed through `exec.Command()` to
        invoke the specified chaincode.'
      parameters:
      - description: channel name (mychannel), cc name (basic), function ('{
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/chaincode.invokeCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Invoke the specified chaincode.
      tags:
      - chaincode
  /fabric/chaincode/query:
    get:
      consumes:
      - application/json
      description: '`peer chaincode invoke` is executed through `exec.Command()` to
        get endorsed result of chaincode function call and print it. It won''t generate
        transaction.'
      parameters:
      - description: channel name (mychannel), cc name (basic)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/chaincode.invokeCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Query using the specified chaincode.
      tags:
      - chaincode
...
  /fabric/lifecycle/admin/{organization}:
    post:
      consumes:
      - application/json
      description: Use terminal environmental variables to set the admin for peer
        cli container. Only Org1 and Org2 are supported.
      parameters:
      - description: organization to be set as admin (Org1 and Org2 supported)
        in: path
        name: organization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lifecycle.currentAdmin'
      summary: Set an org as the admin.
      tags:
      - lifecycle
  /fabric/lifecycle/approve:
    get:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode queryapproved` is executed through `exec.Command()`
        to query approved chaincode definitions.'
      parameters:
      - description: cc name and the channel it was approved in
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.queryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/lifecycle.approvedChaincodeResponse'
      summary: Query an approved chaincode definition on a channel.
      tags:
      - lifecycle
    post:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode approveformyorg` is executed through
        `exec.Command()` to approve a chaincode definition.'
      parameters:
      - description: channel name (mychannel), cc name (basic), cc version (1.0),
          cc sequence (1), package ID (run [GET] /fabric/lifecycle/install)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.approveCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Approve the cc definition for the current org.
      tags:
      - lifecycle
  /fabric/lifecycle/commit:
    get:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode querycommited` is executed through `exec.Command()`
        to query committed chaincode definitions.'
      parameters:
      - description: cc name and the channel it was committed in
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.queryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/lifecycle.committedChaincodeResponse'
      summary: Query the committed chaincode definitions by channel on a peer.
      tags:
      - lifecycle
    post:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode commit` is executed through `exec.Command()`
        to commit chaincode definition on a channel.'
      parameters:
      - description: channel name (mychannel), cc name (basic), cc version (1.0),
          cc sequence (1)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.commitCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Commit the chaincode definition on the channel.
      tags:
      - lifecycle
...
  /fabric/lifecycle/install/{package_name}:
    post:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode install` is executed through `exec.Command()`
        to install chaincode on a peer.'
      parameters:
      - description: name of the package to install (e.g. basic.tar.gz)
        in: path
        name: package_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Install a cc.
      tags:
      - lifecycle
  /fabric/lifecycle/package:
    post:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode install` is executed through `exec.Command()`
        to install chaincode on a peer.'
      parameters:
      - description: name of the cc to package (e.g. asset-transfer-basic), the language
          it is written in, and the label and package name for the cc once packaging
          is done
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.packageCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Package a cc.
      tags:
      - lifecycle
...
swagger: "2.0"